requirements.txt
requests>=2.28.0
PyYAML>=6.0
jsonschema>=4.0.0
neo4j>=5.8.0
fastapi>=0.95.0
uvicorn>=0.22.0
openai>=0.27.0
python-dotenv>=1.0.0
chromadb>=0.4.22
sentence-transformers>=2.2.2
langchain>=0.1.0
langchain-community>=0.0.10
anthropic>=0.18.0
psycopg2-binary>=2.9.0

# Testing dependencies
pytest>=7.4.0
pytest-cov>=4.1.0
pytest-mock>=3.11.1
pytest-asyncio>=0.21.0
pytest-xdist>=3.3.1
httpx>=0.24.0
psutil>=5.9.0
